Estabilización parcial de RequestEntry

+++ Hearts Of Ink/Assets/Scripts/Controller/Multiplayer/ConfigGameController.cs
@@ -126,7 +126,7 @@ public class ConfigGameController : MonoBehaviour
                 FactionId = player.FactionId,
                 MapSocketId = player.MapSocketId,
                 PlayerName = string.Empty,
-                PlayerType = Player.IA.PLAYER
+                PlayerType = (int) Player.IA.PLAYER
             });
         }
